Move all Anchor HOC’s into hoc folder
